having the messages appear newest first
