+++ src/App.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import Form from "./components/Form";
-import Messages from "./Messages";
+import Messages from "./components/Messages";
 
 export const App = () => {
   const [messages, setMessages] = useState([]);
